split out v1 and v2
