+++ src/Components/index.ts
@@ -1,15 +1,15 @@
 export { default as Navigation } from "./Navigation";
-export { default as Dashboard } from "./Dashboard";
-export { default as Movies } from "./Movies";
-export { default as User } from "./User";
-export { default as ContentContainer } from "./ContentContainer";
+export { default as Dashboard } from "./v1/Dashboard";
+export { default as Movies } from "./v1/Movies";
+export { default as User } from "./v1/User";
+export { default as ContentContainer } from "./v1/ContentContainer";
 export { default as Main } from "./Main";
-export { default as MostWatchedPieChart } from "./Chart/MostWatchedPieChart";
-export { default as ActiveUsersBarChart } from "./Chart/ActiveUsersBarChart";
-export { default as ActiveUserDisplay } from "./User/ActiveUserDisplay";
-export { default as AllUsersTable } from "./Table/AllUsersTable";
-export { default as UserMoviesTable } from "./Table/UserMoviesTable";
+export { default as MostWatchedPieChart } from "./v1/Chart/MostWatchedPieChart";
+export { default as ActiveUsersBarChart } from "./v1/Chart/ActiveUsersBarChart";
+export { default as ActiveUserDisplay } from "./v1/User/ActiveUserDisplay";
+export { default as AllUsersTable } from "./v1/Table/AllUsersTable";
+export { default as UserMoviesTable } from "./v1/Table/UserMoviesTable";
 export { default as Header } from "./Header";
-export { default as PieChart } from "./Chart/PieChart";
-export { default as SingleUser } from "./User/SingleUser";
+export { default as PieChart } from "./v1/Chart/PieChart";
+export { default as SingleUser } from "./v1/User/SingleUser";
 export { default as HoverPaper } from "./HoverPaper";
