+++ src/Components/v1/Chart/ActiveUsersBarChart.tsx
@@ -1,7 +1,7 @@
 import { BarChart } from "@mui/x-charts/BarChart";
-import { User } from "../../utils/dummyDataUtils/types";
-import { HoverPaper } from "..";
-import { colors } from "../../utils";
+import { User } from "../../../utils/dummyDataUtils/types";
+import { HoverPaper } from "../..";
+import { colors } from "../../../utils";
 
 const chartSetting = {
   yAxis: [
