+++ src/Components/v1/Dashboard/index.tsx
@@ -1,9 +1,9 @@
 import { Box } from "@mui/material";
-import { useDashboard } from "../../hooks";
-import { ActiveUserDisplay, PieChart, HoverPaper } from "..";
-import { ActiveUsersBarChart } from "..";
-import { User } from "../../utils/dummyDataUtils/types";
-import { colors } from "../../utils";
+import { useDashboard } from "../../../hooks";
+import { ActiveUserDisplay, PieChart, HoverPaper } from "../..";
+import { ActiveUsersBarChart } from "../..";
+import { User } from "../../../utils/dummyDataUtils/types";
+import { colors } from "../../../utils";
 
 const Dashboard = ({
   handleSectionChange,
