+++ src/Components/v1/Table/UserMoviesTable.tsx
@@ -1,34 +1,33 @@
 import { DataGrid, GridColDef } from "@mui/x-data-grid";
 import { Box } from "@mui/material";
-import { UserTable } from "../../utils/dashboardUtils/types";
-import { colors } from "../../utils";
+import { UserMovieTable } from "../../../utils/dashboardUtils/types";
+import { colors } from "../../../utils";
 
 const columns: GridColDef[] = [
-  { field: "id", headerName: "ID", width: 70 },
-  { field: "name", headerName: "Name", width: 130 },
-  { field: "email", headerName: "Email", width: 130 },
-  { field: "username", headerName: "User Name", width: 130 },
-  { field: "active", headerName: "Is Active", width: 130 },
+  { field: "id", headerName: "ID", width: 250 },
+  { field: "title", headerName: "Title", width: 200 },
+  { field: "isFavorite", headerName: "Favorite", width: 100 },
+  { field: "hasEnded", headerName: "Finished", width: 130 },
 ];
 
-const AllUsersTable = ({
-  formatedUsers,
-  handleUserClick,
+const UserMovies = ({
+  formatedMovies,
+  handleUserClick = () => {},
 }: {
-  formatedUsers: UserTable[];
+  formatedMovies: UserMovieTable[];
   handleUserClick: (id: number) => void;
 }) => {
   return (
     <Box
       sx={{
-        height: 650,
+        height: 400,
         width: "100%",
         backgroundColor: colors.base.lightGray,
         cursor: "pointer",
       }}
     >
       <DataGrid
-        rows={formatedUsers}
+        rows={formatedMovies}
         onRowClick={(e) => handleUserClick(e.row.id)}
         columns={columns}
         initialState={{
@@ -42,4 +41,4 @@ const AllUsersTable = ({
   );
 };
 
-export default AllUsersTable;
+export default UserMovies;
