+++ src/Components/v1/User/SingleUser.tsx
@@ -4,9 +4,9 @@ import Box from "@mui/material/Box";
 import Typography from "@mui/material/Typography";
 import Button from "@mui/material/Button";
 import { useNavigate } from "react-router-dom";
-import { formatUserMoviesForTable } from "../../utils";
-import { UserMoviesTable } from "..";
-import { colors } from "../../utils";
+import { formatUserMoviesForTable } from "../../../utils";
+import { UserMoviesTable } from "../..";
+import { colors } from "../../../utils";
 
 const SingleUser = () => {
   const location = useLocation();
