+++ src/Components/v1/User/index.tsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from "react";
-import { AllUsersTable, PieChart } from "..";
-import { useDashboard } from "../../hooks";
-import { formatUsersForTable } from "../../utils";
-import { User } from "../../utils/dummyDataUtils/types";
+import { AllUsersTable, PieChart } from "../..";
+import { useDashboard } from "../../../hooks";
+import { formatUsersForTable } from "../../../utils";
+import { User } from "../../../utils/dummyDataUtils/types";
 import Box from "@mui/material/Box";
 import { useNavigate } from "react-router-dom";
